improve shell/docker/make build/release pipeline

+++ release.sh
@@ -12,10 +12,24 @@ else
 	full_version="${version}-${commit}"
 fi
 
+OS_LIST=(linux darwin)
+
+ARCHIVE=false
+if [ "$1" = "archive" ]; then
+	ARCHIVE=true
+	shift # remove element from arguments
+fi
+
 # Overide the file names to edge version, keep full version at the git commit since
 # that is the expected format
 if [ "$1" = "edge" ]; then
 	version="edge"
+	# if more params defined, use them as OS_LIST
+	if [[ "$#" -gt 1 ]]; then
+		OS_LIST=("${@:2}")
+	fi
+elif [[ "$#" -ge 1 ]]; then
+	OS_LIST=("${@:1}")
 fi
 
 # ensure xgo is installed
@@ -26,17 +40,28 @@ tmpfolder="release/tfchain-xc.tmp"
 rm -rf "$tmpfolder"
 mkdir -p "$tmpfolder"
 
+# Compile targets list
+TARGETS=""
+for os in "${OS_LIST[@]}"; do
+	TARGETS+="${os}/amd64,"
+done
+TARGETS="${TARGETS%?}"
+
 # Compile binaries
 for pkg in ./cmd/tfchainc ./cmd/tfchaind ./cmd/bridged; do
-	xgo --go 1.11.x --targets='linux/amd64,darwin/amd64' \
+	xgo --go 1.11.x --targets="${TARGETS}" \
 		-ldflags="-X ${package}/pkg/config.rawVersion=${full_version} -s -w" \
 		-out "$(basename $pkg)-$version" \
 		-dest "$tmpfolder" \
 		$pkg
 done
 
+if [ "$ARCHIVE" = false ] ; then
+    exit 0 # finished already
+fi
+
 # Create archives
-for os in linux darwin; do
+for os in "${OS_LIST[@]}"; do
 	folder="release/tfchain-${version}-${os}-amd64"
 	rm -rf "$folder"
 	mkdir -p "$folder/cmd"
